Profile: Only allow revising your own notes, limit scope of change for revising

+++ app/controllers/event_notes_controller.rb
@@ -47,6 +47,7 @@ class EventNotesController < ApplicationController
   def update
     safe_params = params.permit(:id, :student_id, event_note: [:text])
     event_note = authorized_or_raise! { EventNote.find(safe_params[:id]) }
+    raise Exceptions::EducatorNotAuthorized unless event_note.educator_id == current_educator.id
 
     # First store the current state of the existing event note
     event_note_revision = create_event_note_revision(event_note)
@@ -54,11 +55,13 @@ class EventNotesController < ApplicationController
       return render json: { errors: event_note_revision.errors.full_messages }, status: 422
     end
 
-    # Update the EventNote
+    # Update the EventNote, but don't update `recorded_at` - keep that at the original
+    # recording time.  The assumption is that these "updates" are from from the user continually
+    # editing during a meeting (the "start time" is fine) or from typo fixes (we don't care about
+    # those).
     safe_params[:event_note]
     if event_note.update({
-      text: safe_params[:event_note][:text],
-      recorded_at: Time.now
+      text: safe_params[:event_note][:text]
     })
       serializer = EventNoteSerializer.dangerously_include_restricted_note_text(event_note)
       render json: serializer.serialize_event_note
@@ -74,6 +77,8 @@ class EventNotesController < ApplicationController
     event_note = authorized_or_raise! do
       EventNoteAttachment.find(id).try(:event_note)
     end
+    raise Exceptions::EducatorNotAuthorized unless event_note.educator_id == current_educator.id
+
     event_note_attachment = event_note.event_note_attachments.find(id)
     if event_note_attachment.destroy
       render json: {}
